+++ Patches/EntranceTeleportPatch.cs
@@ -1,4 +1,6 @@
 ﻿using GameNetcodeStuff;
+using GeneralImprovements.OtherMods;
+using GeneralImprovements.Utilities;
 using HarmonyLib;
 using System.Reflection;
 using UnityEngine;
@@ -19,6 +21,18 @@ namespace GeneralImprovements.Patches
                 }
 
                 return _exitPoint;
+            }
+        }
+
+        [HarmonyPatch(typeof(EntranceTeleport), nameof(Awake))]
+        [HarmonyPostfix]
+        private static void Awake(EntranceTeleport __instance)
+        {
+            // If configured, add a scan node to all exits and fire entrances
+            if (Plugin.ShowDoorsOnScanner.Value && !(__instance.entranceId == 0 && __instance.isEntranceToBuilding))
+            {
+                string text = __instance.isEntranceToBuilding ? "Fire Entrance" : __instance.entranceId == 0 ? "Main Exit" : $"Fire Exit{(MimicsHelper.IsActive ? "?" : "")}";
+                ItemHelper.CreateScanNodeOnObject(__instance.gameObject, 0, 1, __instance.isEntranceToBuilding ? 50 : 20, text);
             }
         }